Basic config structure and implementation
 _____                     ___     ____
 |  ___|   ___     ___     |__ \   | __ )    __ _   _ __
 | |_     / _ \   / _ \      / /   |  _ \   / _` | | '__|
 |  _|   | (_) | | (_) |    |_|    | |_) | | (_| | | |
 |_|      \___/   \___/     (_)    |____/   \__,_| |_|

What is a Foo? What is a Bar?

What would any of them be without someone to configure them?

Implementing a new type of configuration style (cause JSON
has become way to mainstream for the FooBar Boys.

Introducing TOML, because, why not?

Created a new Go module to handle the reading and loading
of the configuration.

API parameters were replaced with values from the TOML
config file.

Signed-off-by: Konstantinos Vytiniotis <konst.vyti@hotmail.com>