New endpoint for section based search
  _____                                __   _   _            _____  _____
 |  __ \                              / _| | | | |          |  __ \|  __ \
 | |  | | __ ___      ___ __     ___ | |_  | |_| |__   ___  | |__) | |__) |
 | |  | |/ _` \ \ /\ / / '_ \   / _ \|  _| | __| '_ \ / _ \ |  ___/|  _  /
 | |__| | (_| |\ V  V /| | | | | (_) | |   | |_| | | |  __/ | |    | | \ \
 |_____/ \__,_| \_/\_/ |_| |_|  \___/|_|    \__|_| |_|\___| |_|    |_|  \_\

Added a new endpoing (/search/section/:request_id).

This endpoint will be iterating over the log files and will
be returning only the log sections related to the
specific request id instead of the whole log file.

TODO: Currenly the endpoint returns the section in string
format. Create a new temp file instead to serve it through
the endpoint.

Signed-off-by: Konstantinos Vytiniotis <konst.vyti@hotmail.com>